College Management System using node js
If you want to keep track of some information about a certain College School, you can use the College Management System. It’s a Node JS and MySQL project. There are three types of roles in this system: the Admin, the Staff, and the People who use it. It’s the person in charge of keeping and managing the data in the system. Users can see some information and join the forum. The Admin User is the person in charge of that job. The system was made for a specific school, but it’s easy to change for your own growth.